cxdemo/backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pymongo>=4.6.0
anthropic>=0.25.0
python-dotenv>=1.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
Werkzeug>=3.0.0
openai>=1.0.0
gunicorn>=21.2.0
